requirements.txt
streamlit
pymongo
streamlit-option-menu
python-dotenv
bcrypt
pymupdf  # (for fitz)
mistralai

